Move componentReplacement fn into create file
