+++ src/create.js
@@ -1,0 +1,6 @@
+const componentReplacement = (content, componentName) =>
+  content
+    .split("${componentToBeTested}")
+    .join(componentName);
+
+export { componentReplacement };
