+++ test/create.test.js
@@ -1,0 +1,12 @@
+import {
+  componentReplacement
+} from "../src/create";
+
+describe("componentReplacement", () => {
+  it("returns the component name from the file", () => {
+    const template = "<${componentToBeTested}>Some</${componentToBeTested}>";
+    const component = "Component";
+    const newTest = `<${component}>Some</${component}>`;
+    expect(componentReplacement(template, component)).toEqual(newTest);
+  });
+});
